refactor: Split method invocation part out to EndpointInvoker
The EndpointInvoker still has knowledge of an HTTP request at this point.
It should be refactored out later.
